refactor: move route handlers to separate module
